flask-backend/requirements.txt
Flask
paho-mqtt
pymongo